# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Aplikasi REST API",
    "description": "Dokumentasi yang dihasilkan dengan <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint untuk mengambil semua User.",
        "responses": {
          "200": {
            "description": "Output API untuk objek User",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint untuk membuat user baru.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Informasi User.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint untuk mengambil User berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi User.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Endpoint untuk menghapus User.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/statistikpendidikan/": {
      "get": {
        "tags": [
          "StatistikPend"
        ],
        "description": "Endpoint untuk mengambil semua StatistikPend.",
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPend",
            "schema": {
              "$ref": "#/definitions/StatistikPends"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "StatistikPend"
        ],
        "description": "Endpoint untuk membuat StatistikPend baru.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Informasi StatistikPend.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatistikPend"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPend",
            "schema": {
              "$ref": "#/definitions/StatistikPend"
            }
          }
        }
      }
    },
    "/statistikpendidikan/{id}": {
      "get": {
        "tags": [
          "StatistikPend"
        ],
        "description": "Endpoint untuk mengambil StatistikPend berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPend",
            "schema": {
              "$ref": "#/definitions/StatistikPend"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "StatistikPend"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi StatistikPend.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatistikPend"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPend",
            "schema": {
              "$ref": "#/definitions/StatistikPend"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StatistikPend"
        ],
        "description": "Endpoint untuk menghapus StatistikPend.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPend",
            "schema": {
              "$ref": "#/definitions/StatistikPend"
            }
          }
        }
      }
    },
    "/statistikpekerjaan/": {
      "get": {
        "tags": [
          "StatistikPekerjaan"
        ],
        "description": "Endpoint untuk mengambil semua StatistikPekerjaan.",
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPekerjaan",
            "schema": {
              "$ref": "#/definitions/StatistikPekerjaans"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "StatistikPekerjaan"
        ],
        "description": "Endpoint untuk membuat StatistikPekerjaan baru.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Informasi StatistikPekerjaan.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatistikPekerjaan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPekerjaan",
            "schema": {
              "$ref": "#/definitions/StatistikPekerjaan"
            }
          }
        }
      }
    },
    "/statistikpekerjaan/{id}": {
      "get": {
        "tags": [
          "StatistikPekerjaan"
        ],
        "description": "Endpoint untuk mengambil StatistikPekerjaan berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPekerjaan",
            "schema": {
              "$ref": "#/definitions/StatistikPekerjaan"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "StatistikPekerjaan"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi StatistikPekerjaan.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatistikPekerjaan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPekerjaan",
            "schema": {
              "$ref": "#/definitions/StatistikPekerjaan"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StatistikPekerjaan"
        ],
        "description": "Endpoint untuk menghapus StatistikPekerjaan.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek StatistikPekerjaan",
            "schema": {
              "$ref": "#/definitions/StatistikPekerjaan"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Iqbal"
        },
        "email": {
          "type": "string",
          "example": "iqbalsaputra@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "12345"
        },
        "notlp": {
          "type": "string",
          "example": "081387887675"
        }
      }
    },
    "Users": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/User"
      }
    },
...
    "StatistikPend": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "sd": {
          "type": "string",
          "example": "30"
        },
        "smp": {
          "type": "string",
          "example": "50"
        },
        "sma": {
          "type": "string",
          "example": "30"
        },
        "smk": {
          "type": "string",
          "example": "50"
        },
        "kuliah": {
          "type": "string",
          "example": "50"
        }
      }
    },
    "StatistikPends": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/StatistikPend"
      }
    },
    "StatistikPekerjaan": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "pns": {
          "type": "string",
          "example": "70"
        },
        "pegawai_swasta": {
          "type": "string",
          "example": "50"
        },
        "wirausaha": {
          "type": "string",
          "example": "30"
        },
        "tni": {
          "type": "string",
          "example": "50"
        },
        "polri": {
          "type": "string",
          "example": "50"
        },
        "buruh": {
          "type": "string",
          "example": "50"
        },
        "tidak_bekerja": {
          "type": "string",
          "example": "50"
        }
      }
    },
    "StatistikPekerjaans": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/StatistikPekerjaan"
      }
    },
...
  }
}